Change rotation order
Rotations must be applied in reverse order because we are rotating the
incoming ray and not the crystal. I changed the rotation order to normal
inside the getRotationMatrix() function, and it will only be inverted
when the ray is multiplied with the matrix. This way it is easier to
make sense of the rotations, since you can think like they are being
applied to the crystal.