Approximate solar limb darkening
Instead of recalculating limb darkening spectrally for every point on
the solar disk, the CIE XYZ coordinates of the top of the solar disk are
now calculated with and without limb darkening, and their ratio is
stored. The ratio is then used to adjust sun color and intensity towards
the edges of the sun disk.