Remove unnecessary spinlock texture
The spinlock texture was meant to prevent raytracing shader invocations
from messing up each others results when they need to write to the same
pixel in the output image.

With correct usage of memory barriers and using the "coherent" keyword
for the output image in the shader, the spinlock code shouldn't be
needed. Based on tentative benchmarking it didn't have much effect on
performance, but it is good to remove it for clarity.