Remove extra normal check for better performance
Finding a ray-triangle intersection tries to break out early if the ray
and triangle normals point the same way. Eventually the the algorithm
will break out anyway in that case, so the initial check is safe to
remove.

Removing the check improves simulation rate by 20% on a GTX 3070 in the
default scenario, probably because breaking out of the loop early
stalls the GPU core.